nanobind/multibody : start exposing joint data
